requirements.txt
django-crispy-forms==1.8.1
Pillow==6.2.1
pkg-resources==0.0.0
